progress

+++ src/main/java/com/lito/fupin/business/organize/OrganizeBusinessService.java
@@ -28,12 +28,32 @@ public class OrganizeBusinessService implements IOrganizeBusinessService {
     @Transactional(rollbackFor = Exception.class)
     @Override
     public Map createOrganize(Map in) throws Exception {
+        String token = in.get("token").toString();
         String organizeName = in.get("organizeName").toString();
         String pid = (String) in.get("pid");
+
+        User user = iUserService.getUserByToken(token);
+        if (user == null) {
+            throw new Exception("10003");
+        }
 
         Organize organize = new Organize();
         organize.setOrganizeId(GGF.UUID().toString());
         organize.setOrganizeName(organizeName);
+        if (pid == null) {
+            /**
+             * 如果pid为空，则赋值为当前用户的organizeid
+             */
+            pid = user.getOrganizeId();
+        } else {
+            /**
+             * 检查pid是否存在
+             */
+            Organize porganize = iOrganizeService.getOrganizeById(pid);
+            if (porganize == null) {
+                throw new Exception("10004");
+            }
+        }
         organize.setPid(pid);
         iOrganizeService.createOrganize(organize);
 
@@ -72,21 +92,49 @@ public class OrganizeBusinessService implements IOrganizeBusinessService {
     @Transactional(rollbackFor = Exception.class)
     @Override
     public void updateOrganize(Map in) throws Exception {
+        String token=in.get("token").toString();
         String organizeId = in.get("organizeId").toString();
         String organizeName = (String) in.get("organizeName");
         String pOrgName = (String) in.get("pOrgName");
+        String pid=(String)in.get("pid");
 
+        /**
+         * 检查用户登录
+         */
+        User user=iUserService.getUserByToken(token);
+        if(user==null){
+            throw new Exception("10003");
+        }
+
+        /**
+         * 读取要修改的机构
+         */
         Organize organize = iOrganizeService.getOrganizeById(organizeId);
         int pp = 0;
         if (!organize.getOrganizeName().equals(organizeName)) {
+            //当前机构名称与要修改的名称不一致，修改
             organize.setOrganizeName(organizeName);
             pp++;
         }
         if (pOrgName != null) {
-            Organize pOrganize = iOrganizeService.getOrganizeByName(pOrgName);
-            organize.setPid(pOrganize.getOrganizeId());
-            pp++;
+            //检查上级机构是否存在
+            if(!organize.getPname().equals(pOrgName)) {
+                Organize pOrganize = iOrganizeService.getOrganizeByName(pOrgName);
+                organize.setPid(pOrganize.getOrganizeId());
+                pp++;
+            }
+        }else{
+            if(pid!=null){
+                //检查上级机构
+                if(!organize.getPid().equals(pid)) {
+                    Organize pOrganize = iOrganizeService.getOrganizeById(pid);
+                    organize.setPid(pid);
+                    pp++;
+                }
+            }
         }
+
+        //当pp>0，则存在可修改的行为
         if (pp > 0) {
             iOrganizeService.updateOrganize(organize);
         }
@@ -94,7 +142,19 @@ public class OrganizeBusinessService implements IOrganizeBusinessService {
 
     @Override
     public void deleteOrganize(Map in) throws Exception {
+        String token=in.get("token").toString();
         String organizeId = in.get("organizeId").toString();
+
+        User user=iUserService.getUserByToken(token);
+        if(user==null){
+            throw new Exception("10003");
+        }
+
+        Organize organize=iOrganizeService.getOrganizeById(organizeId);
+        if(organize==null){
+            throw new Exception("10004");
+        }
+
         iOrganizeService.deleteOrganize(organizeId);
     }
 
@@ -110,7 +170,9 @@ public class OrganizeBusinessService implements IOrganizeBusinessService {
 
         Organize organize = iOrganizeService.getOrganizeById(user.getOrganizeId());
 
-        ArrayList<Organize> organizeList=listSubOrganize(organize);
+        ArrayList<Organize> organizeList = new ArrayList<>();
+        organizeList.add(organize);
+        organizeList.addAll(listSubOrganize(organize));
 
         Map out = new HashMap();
         out.put("organizeList", organizeList);
